requirements.txt
langchain==0.3.21
langchain-community==0.3.20
torch==2.4.1
transformers==4.44.2
python-decouple==3.8
uvicorn==0.34.0
supabase==2.15.0
fastapi==0.115.11
google-generativeai==0.8.4
underthesea==6.8.4